fix(fluid): 微探针去 f64（CI 纪律门"零 f64"抓到的）—— 该门扫 `src/**`，`#[cfg(test)]` 也在扫描面内
上一提交的微探针里用了 `as f64`（累加 sink），触发 CI 的 `静态 (文本/工作流/纪律)` 门：
`discipline_scan.sh` 的 **f64 规则扫 `crates/*/src/**`**，而**白名单只有 `examples/` 与 `tests/`**
——我的探针在 `src/lib.rs` 的 `#[cfg(test)] mod tests` 里 ⇒ **仍在扫描面内**。
改成 f32 累加（`as_secs_f64` 不触发：正则要求 f64 前是非单词字符，`_` 是单词字符 ⇒ 不匹配）。

**本地先把 CI 该 job 的全部门过一遍再推**（避免二次红）：typos 0、vocab ✅、**discipline ✅**、
deps_lock ✅、ci_shape_lock ✅；探针复跑读数不变（收集 22.8 / 算术 37.2 ms）。

+++ crates/vxl-phys-fluid/src/lib.rs
@@ -1996,20 +1996,20 @@ fn micro_probe_gather_vs_compute() {
     let nf = f.n_fluid;
     // ① 收集版
     let t0 = Instant::now();
-    let mut sink = 0.0f64;
+    let mut sink = 0.0f32;
     for i in 0..nf {
         let vi = f.vel[i];
         let rho_i = f.dens[i];
         f.for_neighbors(i, |j, d, r2| {
-            sink += (f.pos[j].x + f.vel[j].y + f.dens[j] + f.press[j] + f.pmass[j]) as f64;
-            sink += (d.x + r2) as f64;
-            sink += (vi.y + rho_i) as f64;
+            sink += f.pos[j].x + f.vel[j].y + f.dens[j] + f.press[j] + f.pmass[j];
+            sink += d.x + r2;
+            sink += vi.y + rho_i;
         });
     }
     let gather = t0.elapsed().as_secs_f64() * 1e3;
     // ② 算术版（真实力相位的算式，但结果丢弃）
     let t0 = Instant::now();
-    let mut sink2 = 0.0f64;
+    let mut sink2 = 0.0f32;
     for i in 0..nf {
         let vi = f.vel[i];
         let rho_i = f.dens[i];
@@ -2036,7 +2036,7 @@ fn micro_probe_gather_vs_compute() {
             let w = k6 * tt * tt * tt;
             xs += (f.vel[j] - vi) * (mj * 2.0 / (rho_i + rho_j) * w);
         });
-        sink2 += (a.x + xs.y) as f64;
+        sink2 += a.x + xs.y;
     }
     let compute = t0.elapsed().as_secs_f64() * 1e3;
     println!(
